ERP/app/requirements.txt
anyio==3.6.2
click==8.1.3
fastapi==0.95.1
gunicorn==20.1.0
h11==0.14.0
idna==3.4
pydantic==1.10.7
sniffio==1.3.0
starlette==0.26.1
typing_extensions==4.5.0
uvicorn==0.21.1
python-dotenv==1.0.0
SQLAlchemy==1.4.31
PyMySQL==1.0.2
pandas==1.5.3
